+++ src/main/java/subway/repository/StationLineRepository.java
@@ -1,6 +1,7 @@
-package subway.stationline;
+package subway.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.entity.StationLine;
 
 public interface StationLineRepository extends JpaRepository<StationLine, Long> {
