+++ src/main/java/subway/service/StationService.java
@@ -1,10 +1,14 @@
-package subway.station;
+package subway.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import subway.entity.Station;
+import subway.repository.StationRepository;
+import subway.service.request.StationRequest;
+import subway.service.response.StationResponse;
 
 @Service
 @Transactional(readOnly = true)
