+++ src/main/java/subway/service/request/StationLineRequest.java
@@ -1,4 +1,4 @@
-package subway.stationline;
+package subway.service.request;
 
 public class StationLineRequest {
