+++ src/main/java/subway/service/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.service.request;
 
 public class StationRequest {
     private String name;
