+++ src/test/java/subway/StationAcceptanceTest.java
@@ -14,35 +14,40 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.station.StationResponse;
+import subway.service.response.StationResponse;
 
 @DisplayName("지하철역 관련 기능")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 public class StationAcceptanceTest {
+
+    private static final String 지하철역이름 = "지하철역이름";
+    private static final String 새로운지하철역이름 = "새로운지하철역이름";
+    private static final String 또다른지하철역이 = "또다른지하철역이";
+    private static final String 강남역 = "강남역";
+
+
     /**
-     * When 지하철역을 생성하면
-     * Then 지하철역이 생성된다
-     * Then 지하철역 목록 조회 시 생성한 역을 찾을 수 있다
+     * When 지하철역을 생성하면 Then 지하철역이 생성된다 Then 지하철역 목록 조회 시 생성한 역을 찾을 수 있다
      */
     @DisplayName("지하철역을 생성한다.")
     @Test
     void createStation() {
         // when
-        ExtractableResponse<Response> response = createStation("강남역");
+        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역);
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
 
         // then
         List<String> stationNames =
-                RestAssured.given().log().all()
-                        .when().get("/stations")
-                        .then().log().all()
-                        .extract().jsonPath().getList("name", String.class);
-        assertThat(stationNames).containsAnyOf("강남역");
+            RestAssured.given().log().all()
+                .when().get("/stations")
+                .then().log().all()
+                .extract().jsonPath().getList("name", String.class);
+        assertThat(stationNames).containsAnyOf(강남역);
     }
 
-    private static ExtractableResponse<Response> createStation(final String name) {
+    private static ExtractableResponse<Response> 지하철역_생성_요청(final String name) {
 
         Map<String, String> params = new HashMap<>();
 
@@ -57,37 +62,31 @@ public class StationAcceptanceTest {
     }
 
     /**
-     * Given 2개의 지하철역을 생성하고
-     * When 지하철역 목록을 조회하면
-     * Then 2개의 지하철역을 응답 받는다
+     * Given 2개의 지하철역을 생성하고 When 지하철역 목록을 조회하면 Then 2개의 지하철역을 응답 받는다
      */
     @Test
     @DisplayName("지하철역 목록을 조회한다.")
     void showStations() {
 
         //given
-        String name1 = "지하철역이름";
-        String name2 = "새로운지하철역이름";
-        String name3 = "또다른지하철역이";
-
-        createStation(name1);
-        createStation(name2);
-        createStation(name3);
+        지하철역_생성_요청(지하철역이름);
+        지하철역_생성_요청(새로운지하철역이름);
+        지하철역_생성_요청(또다른지하철역이);
 
         //when
-        ExtractableResponse<Response> response = getStationsList();
+        ExtractableResponse<Response> response = 지하철역_리스트_조회();
 
         //then
         Assertions.assertAll(
             () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
             () -> assertThat(response.jsonPath().getList("", StationResponse.class)).hasSize(3),
-            () -> assertThat(response.jsonPath().getString("[0].name")).isEqualTo(name1),
-            () -> assertThat(response.jsonPath().getString("[1].name")).isEqualTo(name2),
-            () -> assertThat(response.jsonPath().getString("[2].name")).isEqualTo(name3)
+            () -> assertThat(response.jsonPath().getString("[0].name")).isEqualTo(지하철역이름),
+            () -> assertThat(response.jsonPath().getString("[1].name")).isEqualTo(새로운지하철역이름),
+            () -> assertThat(response.jsonPath().getString("[2].name")).isEqualTo(또다른지하철역이)
         );
     }
 
-    private static ExtractableResponse<Response> getStationsList() {
+    private static ExtractableResponse<Response> 지하철역_리스트_조회() {
         return RestAssured
             .given().log().all()
             .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -97,16 +96,14 @@ public class StationAcceptanceTest {
     }
 
     /**
-     * Given 지하철역을 생성하고
-     * When 그 지하철역을 삭제하면
-     * Then 그 지하철역 목록 조회 시 생성한 역을 찾을 수 없다
+     * Given 지하철역을 생성하고 When 그 지하철역을 삭제하면 Then 그 지하철역 목록 조회 시 생성한 역을 찾을 수 없다
      */
     @Test
     @DisplayName("지하철역 삭제")
     void removeStation() {
 
         //given
-        ExtractableResponse<Response> createResponse = createStation("새로운지하철역");
+        ExtractableResponse<Response> createResponse = 지하철역_생성_요청("새로운지하철역");
 
         //when
         RestAssured
@@ -118,7 +115,7 @@ public class StationAcceptanceTest {
             .extract();
 
         //then
-        ExtractableResponse<Response> stationsList = getStationsList();
+        ExtractableResponse<Response> stationsList = 지하철역_리스트_조회();
 
         //then
         assertThat(stationsList.jsonPath().getList("")).isEmpty();
